- downloading classifications from API - checking to see if grid/annotation files exists already before trying to download them

+++ api.py
@@ -5,11 +5,11 @@ import urllib
 import zipfile
 
 API_HOST = 'http://api.brain-map.org/'
-API_QUERY_BASE_URL = API_HOST + 'api/v2/data/query.json?'
+API_QUERY_BASE_URL = API_HOST + 'api/v2/data/query.json'
 GRID_URL_FORMAT = API_HOST + "grid_data/download/%d"
 
 # Make a query to the api, downloading all rows and packaging them into an array.
-def query(url, rows_per_query=2000):
+def query(query_string, rows_per_query=2000):
     start_row = 0
     total_rows = -1
     
@@ -21,12 +21,11 @@ def query(url, rows_per_query=2000):
     while not done:
 
         # Append start_row and num_row arguments to the query.
-        paged_url = '%s%s,rma::options[start_row$eq%d][num_rows$eq%d]' % (API_QUERY_BASE_URL, url, start_row, rows_per_query)
-        
-        print paged_url
+        data = { 'criteria': '%s,rma::options[start_row$eq%d][num_rows$eq%d]' % (query_string, start_row, rows_per_query) }
+        print data
 
         # Convert the response to JSON
-        response = urllib.urlopen(paged_url).read()
+        response = urllib.urlopen(API_QUERY_BASE_URL, urllib.urlencode(data)).read()
         response_json = json.loads(response)
 
         if not response_json['success']:
@@ -49,17 +48,21 @@ def query(url, rows_per_query=2000):
 # Download a zip file containing the grid data for a section data set.
 # Return the results as an array.
 def download_grid_file(section_data_set_id, file_prefix=''):
+    # Decide what to call the mhd/raw when we save them.
+    energy_mhd_file_name = '%s%d_energy.mhd' % (file_prefix, section_data_set_id)
+    energy_raw_file_name = '%s%d_energy.raw' % (file_prefix, section_data_set_id)
+
+    if os.path.exists(energy_raw_file_name) and os.path.exists(energy_raw_file_name):
+        return os.path.basename(energy_mhd_file_name), os.path.basename(energy_raw_file_name)
+
     url = GRID_URL_FORMAT % (section_data_set_id)
+    print url
     fh = urllib.urlretrieve(url)
 
     zf = zipfile.ZipFile(fh[0])
     
     header = zf.read('energy.mhd')
     raw = zf.read('energy.raw')
-
-    # Decide what to call the mhd/raw when we save them.
-    energy_mhd_file_name = '%s%d_energy.mhd' % (file_prefix, section_data_set_id)
-    energy_raw_file_name = '%s%d_energy.raw' % (file_prefix, section_data_set_id)
 
     # Update the mhd to use the new raw file name.
     header = header.replace('energy.raw', os.path.basename(energy_raw_file_name))
@@ -73,26 +76,37 @@ def download_grid_file(section_data_set_id, file_prefix=''):
     return os.path.basename(energy_mhd_file_name), os.path.basename(energy_raw_file_name)
 
 # Download all of the section data sets from a particular product
-def download_data_sets(product_id):
-    rows = query("criteria=model::SectionDataSet,rma::criteria,[failed$eq'false'],[expression$eq'true'],products[id$eq%s],rma::include,probes,genes,reference_space" % product_id)
-    return rows[0:10]
+def download_data_sets(product_id, reference_space_age_names, plane_of_section_id):
+    age_str = ','.join(("'%s'" % age_name) for age_name in reference_space_age_names)
+    results = query(("model::SectionDataSet" +
+                     ",rma::criteria,[failed$eq'false'],[expression$eq'true'],[plane_of_section_id$eq%d],[storage_directory$ne'']" +
+                     ",reference_space(age[name$in%s]),products[id$eq%s]" +
+                     ",rma::include,probes,genes,reference_space") % (plane_of_section_id, age_str, product_id))
+    return results
 
 # Download all of the structures in a graph
 def download_structures(graph_id):
-    rows = query("criteria=model::Structure,rma::criteria,[graph_id$eq%d]" % (graph_id))
-    return rows
+    results = query("model::Structure,rma::criteria,[graph_id$eq%d]" % (graph_id))
+    return results
 
 # Download the annotated volume file for a gien reference space.
-def download_annotation_volume(reference_space_id, file_prefix):
+def download_annotation_volume(reference_space_id, file_prefix, tmp_file="tmp.zip"):
 
+    # Decide what to call the mhd/raw when we save them.
+    annot_mhd_file_name = '%s%d_annotation.mhd' % (file_prefix, reference_space_id)
+    annot_raw_file_name = '%s%d_annotation.raw' % (file_prefix, reference_space_id)
+
+    if os.path.exists(annot_raw_file_name) and os.path.exists(annot_mhd_file_name):
+        return os.path.basename(annot_mhd_file_name), os.path.basename(annot_raw_file_name)
+    
     # Each reference space has a well known file called 'gridAnnotation.zip'. 
     # Query the API for the link to that file.
-    rows = query("criteria=model::ReferenceSpace,rma::criteria,[id$eq%d],rma::include,well_known_files[path$li'*gridAnnotation.zip']" % reference_space_id)
-    refspace = rows[0]
+    results = query("model::ReferenceSpace,rma::criteria,[id$eq%d],rma::include,well_known_files[path$li'*gridAnnotation.zip']" % reference_space_id)
+    refspace = results[0]
     reffile = refspace['well_known_files'][0]
 
     # Download the zip file.
-    fh = urllib.urlretrieve(API_HOST + reffile["download_link"])
+    fh = urllib.urlretrieve(API_HOST + reffile["download_link"], tmp_file)
     zf = zipfile.ZipFile(fh[0])
 
     # Unzip it and pull out the header and raw file.  Due to inconsistent naming conventions, this could
@@ -110,9 +124,6 @@ def download_annotation_volume(reference_space_id, file_prefix):
     assert header, "Failed to unzip grid annotation header file"
     assert raw, "Failed to unzip grid annotation raw file"
 
-    # Decide what to call the mhd/raw when we save them.
-    annot_mhd_file_name = '%s%d_annotation.mhd' % (file_prefix, reference_space_id)
-    annot_raw_file_name = '%s%d_annotation.raw' % (file_prefix, reference_space_id)
 
     # Update the mhd to use the new raw file name.
     header = header.replace('gridAnnotation.raw', os.path.basename(annot_raw_file_name))
@@ -124,4 +135,18 @@ def download_annotation_volume(reference_space_id, file_prefix):
     with open(annot_raw_file_name, 'wb') as f:
         f.write(raw)
 
+    os.remove(tmp_file)
+
     return os.path.basename(annot_mhd_file_name), os.path.basename(annot_raw_file_name)
+
+def download_gene_classifications(gene_ids):
+    gene_id_str = ",".join(str(gid) for gid in gene_ids)
+    results = query("model::GeneClassification,rma::include,genes[id$in%s]" % gene_id_str)
+
+    classifications = { gene_id: [] for gene_id in gene_ids }
+
+    for classification in results:
+        for gene in classification['genes']:
+            classifications[gene['id']].append(classification['name'])
+    
+    return classifications
